Spread out some structs in more files
