+++ src/ast/mod.rs
@@ -16,6 +16,8 @@ mod head;
 mod identifier;
 mod literal;
 mod module;
+mod op_arity;
+mod op_key;
 mod operator;
 mod pattern;
 mod procession;
@@ -40,7 +42,9 @@ pub(crate) use head::Head;
 pub(crate) use identifier::Identifier;
 pub(crate) use literal::Literal;
 pub(crate) use module::Module;
-pub(crate) use operator::{OpArity, OpKey, Operator};
+pub(crate) use op_arity::OpArity;
+pub(crate) use op_key::OpKey;
+pub(crate) use operator::Operator;
 pub(crate) use pattern::Pattern;
 pub(crate) use procession::Procession;
 pub(crate) use query::Query;
